AI_Project/Youtube_Video_Analyzer/requirements.txt
streamlit
youtube-transcript-api
google-generativeai
python-dotenv
